- Updated installer to latest build version - Fixed underflow issue in MD routines - Replace random number generator with mersenne twister (threadsafe) - Fixed MD routines not working for normalization correction - Small graphing fix for consistent color - Updated to x86 to prevent crashes on Vista 64 bit - Cleaned up Bin directory
git-svn-id: svn://svn.code.sf.net/p/stochfit/code/trunk@46 a99bd192-a644-0410-9c81-d981d67978eb

+++ StochFitMain/StochFitdll/stdafx.h
@@ -59,4 +59,4 @@ using namespace std;
  
 //Use the error function from libmmd.lib (Intel math library), can't use mathimf.h due to
 //compiler bug - hopefully this will be fixed...
-extern "C" double __declspec(dllimport) __cdecl erf(double __x);
+extern "C" float __declspec(dllimport) __cdecl erff(float __x);
